Move tmp views and streamline local testing
