+++ .gitignore
@@ -4,3 +4,4 @@ Homestead.json
 Homestead.yaml
 .env
 composer.phar
+/database/testing.sqlite
