+++ phpunit.xml
@@ -24,6 +24,7 @@
     </filter>
     <php>
         <env name="APP_ENV" value="testing"/>
+        <env name="DB_CONNECTION" value="test"/>
         <env name="CACHE_DRIVER" value="array"/>
     </php>
 </phpunit>
